sfixes

+++ Models/WatchListResponse.cs
@@ -6,7 +6,7 @@ namespace MovieApi.Models
 		public int id { get; set; }
         public string title { get; set; }
 		public string poster_path { get; set; }
-		public double vote_average { get; set; }
+		public int vote_average { get; set; }
     }
 }
